supports double quotes in the alias
big improvement. before, if a user had double quotes in their alias, it would destroy the rest of the $allaliases loop. instead of splitting on the first set of double quotes, we now index the first and last occurrence per line, to get the true alias.

this should make it easier to find/replace special characters moving forward